refactor: share structured SQL console primitives
